feat(memory_watcher): untested fixed heap
This commit relieas the armMbed implementation of malloc and free and proposes a
fixed hep implementation for rt-bench benchmark. The user will have a -H
parameter to specifi where the heap hould start. mallocs will be redirected
there via a custom implementation of sbrk and an mmap of /dev/mem.

Signed-off-by: Mattia Nicolella <mattianicolella@gmail.com>