feat(generator/memory_watcher): support files for heap
This commit adds the possibility to choose a fixed size heap that is being
backed by a file different from /dev/mem.

We do not support yet the possibility to specify anb offset in said file if it's
not /dev/mem.

For both types of files a maximum heap size has to be defined (due to us using
mmap() under the hood).

Signed-off-by: Mattia Nicolella <mattianicolella@gmail.com>