feat(memory_watcher): add option to change aligment for the memory watcher
WIP commit to support fixed heap sizes that are not byte-addressable
Use the new optional features FEAT_MEM_WATCHER_ALIGN=(y/n) to enable it
and MEM_WATCHER_ALIGN=(8/16/32/64) to set an alignment.

This changes the types in memory_watcher_config from void* to uintX_t, so that
pointer arithmetic in __wrap_sbrk should consider the alignment.