refactor: make improvements
This commits enables benchmark suites to reuse existing object files when
possible due to smarter make rules, however it now discouraged to pass variables
to make via the commandline.

All benchmark suites and documentation have been updated to reflect the below
breaking changes and are already using he new variables.

BREAKING_CHANGES:
all BASE_* variables have been renamed to RTBENCH_*

the rtbench  target was renamed to RTBENCH and it is now a variable, so make can
keep track of all the dependencies and do partial recompilation.

make might not correctly trigger a full recompilation if flags are give through
the command line, so edits to the feature flags and make variables should be do
in options.mk (which is now part of the RT-Bench core compilation dependencies)